# docs/architecture/v3/openapi-v3.yaml
openapi: 3.1.0
info:
  title: GDA Command V3 API
  version: 3.0.0-draft
  description: |
    V3 API contract for GDA Command — Envision-only, single-tenant.
    Every list endpoint returns R1 source siblings on every field.
    Every detail endpoint auto-runs R2 analysis.

    **Binding scope:** Envision-only. No `ou_tag` query params.
    Partners are lookup-only (not browsable).
  contact:
    name: GDA Engineering
    email: shawn@envision-is.com
  license:
    name: Proprietary
    identifier: LicenseRef-GDA-Proprietary
servers:
  - url: https://gda.csr-llc.tech/api/v3
    description: Production
  - url: http://localhost:3001/api/v3
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /fast-track:
    post:
      operationId: triageFastTrack
      summary: Triage an opportunity via Fast Track
      description: |
        Computes input_hash, checks cache, enqueues if miss, sync-waits up to 10s.
        Returns 200 with result or 503 ANALYSIS_TIMEOUT.
      tags: [Fast Track]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FastTrackTriageInput'
      responses:
        '200':
          description: Triage result (cache hit or fresh)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FastTrackTriageResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          description: Analysis timed out after 10s sync window
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

    get:
      operationId: listFastTrack
      summary: List recent fast track assessments
      description: |
        Returns newest-first, cursor-paginated list of assessments.
      tags: [Fast Track]
      parameters:
        - name: since
          in: query
          description: ISO-8601 date filter (default 7 days ago)
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          description: Page size (1-100, default 25)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - name: cursor
          in: query
          description: Opaque cursor from previous response
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of assessments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FastTrackListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /fast-track/{id}:
    get:
      operationId: getFastTrack
      summary: Get a single fast track assessment
      tags: [Fast Track]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Single assessment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FastTrackTriageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT bearer token (preserved from V2)

    WebhookKeyAuth:
      type: apiKey
      in: header
      name: x-gda-key
      description: Shared webhook secret from GDA_WEBHOOK_KEY env var

  # -------------------------------------------------------------------------
  # Parameters
  # -------------------------------------------------------------------------
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: "title is required"
              detail: null
            meta:
              generatedAt: "2026-05-29T16:00:00.000Z"
              requestId: req_err001
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: "Missing or invalid authorization"
              detail: null
            meta:
              generatedAt: "2026-05-29T16:00:00.000Z"
              requestId: req_err002
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: "Resource not found"
              detail: null
            meta:
              generatedAt: "2026-05-29T16:00:00.000Z"
              requestId: req_err003
...
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          const: true
        data:
          description: Endpoint-specific payload
        meta:
          $ref: '#/components/schemas/Meta'
    ErrorEnvelope:
      type: object
      required: [success, error, meta]
      properties:
        success:
          type: boolean
          const: false
        error:
          $ref: '#/components/schemas/ApiError'
        meta:
          $ref: '#/components/schemas/Meta'
    Meta:
      type: object
      required: [generatedAt]
      properties:
        generatedAt:
          type: string
          format: date-time
        source:
          type: string
          default: v3
        requestId:
          type: string
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        detail:
          type: string
          nullable: true

    # -- Source citation (R1) -----------------------------------------------
...
    SourceChip:
      type: object
      required: [label, url, kind, retrieved_at]
      properties:
        label:
          type: string
        url:
          type: string
          format: uri
        kind:
          type: string
        retrieved_at:
          type: string
          format: date-time

    # Fast Track schemas (F-227)
    FastTrackTriageInput:
      type: object
      required: [title, description, naics_codes]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        description:
          type: string
          minLength: 1
          maxLength: 50000
        naics_codes:
          type: array
          maxItems: 10
          items:
            type: string
            pattern: '^\d{6}$'
        set_aside:
          type: string
          maxLength: 200
          nullable: true
        place_of_performance:
          type: string
          maxLength: 200
          nullable: true
    FastTrackAssessment:
      type: object
      required: [id, grade, rationale, naics_match_score, recommended_action, source_chips, model_used, generated_at, cache_hit]
      properties:
        id:
          type: string
        grade:
          type: string
          enum: ['A', 'B', 'C']
        rationale:
          type: string
        naics_match_score:
          type: number
          minimum: 0
          maximum: 100
        recommended_action:
          type: string
          enum: [pursue, watch, skip]
        source_chips:
          type: array
          items:
            $ref: '#/components/schemas/SourceChip'
        model_used:
          type: string
        generated_at:
          type: string
          format: date-time
        cache_hit:
          type: boolean
    FastTrackTriageResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/FastTrackAssessment'
    FastTrackListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              type: object
              required: [items, next_cursor]
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/FastTrackAssessment'
                next_cursor:
                  type: string
                  nullable: true